+++ src/api/resetMovement.ts
@@ -1,5 +1,6 @@
-import { MODULE_ID } from "../constants";
+import { LEGACY_MODULE_ID, MODULE_ID } from "../constants";
 
 export function callResetMovement(actor: Actor) {
     Hooks.callAll(`${MODULE_ID}.resetMovement`, actor);
+    Hooks.callAll(`${LEGACY_MODULE_ID}.resetMovement`, actor);
 }